Create BasePaymentSheetActivity
